MDEV-23291: SUM column from a derived table returns invalid values
The issue here was the read_set bitmap was not set for a field which was used as a reference
in an inner select. We need to make sure that if we are in an inner select and we have
references from outer select then we update the table bitmaps for such references.
Introduced a function in the class Item_subselect that would update bitmaps of table for
the references within a subquery that are defined in outer selects.